Move restore.py directly under quickmpc
